requirements.txt
Flask>=2.3,<4
python-dotenv>=1.0,<2
google-generativeai>=0.8,<1
PyMuPDF>=1.24,<1.26
python-dateutil>=2.8,<3
Pillow>=10,<12
pytesseract>=0.3.10,<0.3.13
SQLAlchemy>=2.0,<3
psycopg2-binary>=2.9,<2.11
pytest>=7.4,<9
